% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accounts_2019_2023}
\alias{accounts_2019_2023}
\title{Budget accounts from 2019 to 2023}
\format{
A data frame with 1,042 rows and 7 variables:
\describe{
\item{account.name}{Account name in Tyler}
\item{accountnum}{Combined account number and account description; distinct from account.name in some instances}
\item{accountnum_desc}{Description of account}
\item{accountnum_num}{Number of account}
\item{department}{A department that uses the account}
\item{division}{A division that uses the account}
\item{accountnum_desc_small}{Small-case accountnum_desc}
}
}
\source{
{Tyler budgeting system}
}
\usage{
accounts_2019_2023
}
\description{
Names and numbers used by each Department and Division for all accounts used in the FY 2019 to FY 2023 budget. Each row is a unique combination of the account.name, department, and division.
}
\keyword{datasets}
