% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_prep.R
\name{add_ws_data}
\alias{add_ws_data}
\title{Add worksheet data}
\usage{
add_ws_data(budget_wb, df, sheetname, .newwb = F)
}
\arguments{
\item{budget_wb}{An openxlsx Excel workbook}

\item{df}{Budget dataframe to add to workbook}

\item{sheetname}{Name of sheet to add budget dataframe to as string}

\item{.newwb}{Whether to return the budget workbook at the end of the function; defaults F, just modifying the existing workbook but returning nothing}
}
\value{
Function doesn't return anything unless .newwb = T, but adds a worksheet to an existing openxlsx workbook with formatted data
}
\description{
Adds a budget-dataframe to a sheet of an openxlsx Excel workbook, and formats sheet for output
}
