% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{base_process_extract}
\alias{base_process_extract}
\title{Base-process Tyler Report Extract
Base-processing of unprocessed Tyler report extract dataset. Standardizes column names  for other functions, associates Tyler categories with dataset, makes sure merged successfully}
\usage{
base_process_extract(df)
}
\arguments{
\item{df}{Unprocessed Tyler-report extract}
}
\value{
Base-processed Tyler report extract
}
\description{
Base-process Tyler Report Extract
Base-processing of unprocessed Tyler report extract dataset. Standardizes column names  for other functions, associates Tyler categories with dataset, makes sure merged successfully
}
