% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{budget_cat_df}
\alias{budget_cat_df}
\title{Budget categories documented by Finance in 2021}
\format{
A data frame with 147 rows and 4 variables:
\describe{
\item{budget_cat_overall_num }{Account name in Tyler}
\item{budget_cat_overall_desc}{Combined account number and account description; distinct from account.name in some instances}
\item{accountnum_num}{An account number associated with }
\item{budget_cat_overall}{A category used in the City's budget tables}
}

A data frame with 147 rows and 4 variables:
\describe{
\item{budget_cat_overall_num }{Number of budget category used in City tables}
\item{budget_cat_overall_desc}{Description of budget category used in City tables}
\item{accountnum_num}{Account number associated with budget categories}
\item{budget_cat_overall}{Combination of name and number associated with budget categories used in City tables}
}
}
\source{
{Finance Department}

{Finance Department}
}
\usage{
budget_cat_df

budget_cat_df
}
\description{
Names and numbers of budget account categories used by finance department in FY 2021 budget, documented in PDF.

Names and numbers of budget account categories used by finance department in FY 2021 budget, documented in PDF. Each row is a unique accountnum_num. Note: should use table_accounts_complete for full-association of budget categories with account names and numbers.
}
\keyword{datasets}
