% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{budget_group}
\alias{budget_group}
\title{Group Budget Data}
\usage{
budget_group(df, colsgrp)
}
\arguments{
\item{df}{Tyler Extract Report budget dataset}

\item{colsgrp}{String vector of columns to group the Tyler Extract Report by. For columns that have been separated by tpbudget::sep_budget, only need to identify the basename of the column. For instance, if there are columns for accountnum, accountnum_desc, and accountnum_num, only need to provide "accountnum" to still keep the "accountnum_desc" and "accountnum_num" columns in the group. Note that if there are "change" or "difference" year-columns in the dataset, percentages will still be summed and these should be recalculated}
}
\value{
A dataframe grouped by the provided colsgroup vector, with year-columns summed by the grouped columns
}
\description{
Groups Tyler budget extract report budget data by provided set of columns, summing year columns in the dataset. Year columns in the data should have standard names that will be picked up by tpbudget::pull_year_cols()
}
