% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{calc_yearcol}
\alias{calc_yearcol}
\title{Calculate Change and Differences in Years}
\usage{
calc_yearcol(df, curr_fy = 2022)
}
\arguments{
\item{df}{Tyler Extract Report dataset processed by base_report_extract()}

\item{curr_fy}{Current fiscal year}
}
\value{
Dataframe with columns for
}
\description{
From a base-processed Tyler Extract Report dataset, calculates change in year-to-year values for a given fiscal year, and differences between adopted and projected expenditures for a given fiscal year
}
