% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_prep.R
\name{dept_exp_tables}
\alias{dept_exp_tables}
\title{Create department expenditure tables}
\usage{
dept_exp_tables(dept_filter, bud_dept_list, fy = 2023)
}
\arguments{
\item{dept_filter}{Department number}

\item{bud_dept_list}{List of budget departmental expenditure tables generated by tpbudget::budget_list().}

\item{fy}{Fiscal year of this budget book; default 2023.}
}
\value{
Excel workbook populated with different sheets reflecting different levels of budget data
}
\description{
Creates Excel workbook of standard Departmental expenditure tables for purpose of sharing with departments
}
