% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dept_list}
\alias{dept_list}
\title{Department List}
\format{
A data frame with 12 rows and 3 variables:
\describe{
\item{department}{Department name and number}
\item{department_desc}{Department description}
\item{department_num}{Department number}
}
}
\source{
{Tyler budgeting system}
}
\usage{
dept_list
}
\description{
List of all Departments and their Department-number in Tyler. Each row is a unique department.
}
\keyword{datasets}
