% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_preparation.R
\name{dept_pull}
\alias{dept_pull}
\title{Pull Department name from department code}
\usage{
dept_pull(dc)
}
\arguments{
\item{dc}{Department code}
}
\value{
Department name
}
\description{
Pulls department name from supplied department code
}
