% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_prep.R
\name{df_all_op_funds}
\alias{df_all_op_funds}
\title{All Operating Funds Dataframe}
\usage{
df_all_op_funds(df_rev, df_exp, .fundcol = fund_recode)
}
\arguments{
\item{df_rev}{Base-revenue dataframe; bud_extract_rev}

\item{df_exp}{Base-expenditures dataframe; bud_extract_dept}

\item{.fundcol}{Column corresponding to recoded fund values; default fund_recode}
}
\value{

}
\description{
Function to make operating dataframe for visualization of all operating funds, merging budget and revenue data into one file
}
