% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_prep.R
\name{df_budg_rec_table}
\alias{df_budg_rec_table}
\title{Calculate Overall Revenue}
\usage{
df_budg_rec_table(df_rev, pivot = F)
}
\arguments{
\item{df_rev}{Revenue dataframe}

\item{pivot}{Whether to pivot dataframe into longer format}
}
\value{
Grouped revenue dataframe
}
\description{
Calculates overall revenue in budget-revenue dataset and change by year
}
