% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_prep.R
\name{df_fundtable}
\alias{df_fundtable}
\title{All Operating Funds Table}
\usage{
df_fundtable(rev_df, exp_df, fundac, fy = 2023, fundcol = fund_recode, ...)
}
\arguments{
\item{rev_df}{Base-processed revenue dataframe}

\item{exp_df}{Base-processed expenditures dataframe}

\item{fundac}{Acronym of fund to filter data to.}

\item{fy}{Fiscal year of this budget book; default 2023}

\item{fundcol}{Data-masked name of fund column to supply df_all_op_funds function; default fund_recode}

\item{...}{Optional additional arguments to find_yrstooearly, filter_fund, and revenue_factorize functions.}
}
\value{
Dataframe of all operating funds with revenue and expenditures in one dataframe, filtered to one fund
}
\description{
Generate all operating funds dataframe with combined revenues and expenditures for a fund-filtered to
}
