% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_prep.R
\name{df_propadpt_chng}
\alias{df_propadpt_chng}
\title{Title}
\usage{
df_propadpt_chng(df_exp, fund_ac = "GF", filternot0 = F, .currfy = 2022)
}
\arguments{
\item{df_exp}{Base-processed expendiures dataset}

\item{fund_ac}{Fund-acronym to filter dataframe to; default "GF" for general fund}

\item{filternot0}{Whether to filter dataframe to rows where the difference between the adopted and proposed budget is not 0; default false}

\item{.currfy}{Fiscal-year for the concluding budget (not the fiscal year for the budget being prepared); default 2022}
}
\value{
Dataframe summarized to fund-level expenditures and revenues
}
\description{
function to generate base df for changes to proposed/adopted budget table
}
