% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{df_rename_yeartoyear}
\alias{df_rename_yeartoyear}
\title{Rename year-to-year columns in dataframe
Renames year-to-year change and difference between adopted and adjusted columns in a dataframe.}
\usage{
df_rename_yeartoyear(df)
}
\arguments{
\item{df}{Dataframe with standard year to year change and difference between adopted and adjusted columns.}
}
\value{
Dataframe with year-to-year and difference columns renamed
}
\description{
Rename year-to-year columns in dataframe
Renames year-to-year change and difference between adopted and adjusted columns in a dataframe.
}
