% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_prep.R
\name{df_rev_exp_chngfund}
\alias{df_rev_exp_chngfund}
\title{Generate Combined Statement of Revenue}
\usage{
df_rev_exp_chngfund(df_rev, df_exp)
}
\arguments{
\item{df_rev}{Base-processed budget revenue data.}

\item{df_exp}{Base-processed budget expenditure data.}
}
\value{
Combined dataframe of fund-summarized revenues and expenditures.
}
\description{
Generates dataframe for "Combined Statement of Revenue, Expenditures, & Change in Fund Balances - All Governmental Funds" table in budget book.
}
