% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_prep.R
\name{dir_budg}
\alias{dir_budg}
\title{Create Budget Directory
Creates directory for storing some type of budget data in a data/output/ directory}
\usage{
dir_budg(filetype, dir_string = NULL)
}
\arguments{
\item{filetype}{String name of directory that dir_string directory will be stored in}

\item{dir_string}{String name of sub-directory to create in filetype directory. Default null if just want to create one new directory in data/output/}
}
\value{
Subdirectory (and/or parent directory) in data/output directory
}
\description{
Create Budget Directory
Creates directory for storing some type of budget data in a data/output/ directory
}
\examples{
{
dir_budg("parent_dir", "sub_dir")

dir_budg("parent_dir")

}
}
