% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{div_list}
\alias{div_list}
\title{Division List}
\format{
A data frame with 49 rows and 3 variables:
\describe{
\item{division}{division name and number}
\item{division_desc}{division description}
\item{division_num}{division number}
}
}
\source{
{Tyler budgeting system}
}
\usage{
div_list
}
\description{
List of all Divisions and their Division-number in Tyler. Each row is a unique division.
}
\keyword{datasets}
