% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{div_overall}
\alias{div_overall}
\title{Generate Division Expenditures Overall dataframe}
\usage{
div_overall(divcode, exp_df, fy = 2023)
}
\arguments{
\item{divcode}{Division code}

\item{exp_df}{Base process Tyler expenditures dataframe}

\item{fy}{Fiscal year preparing budget book for; default 2023}
}
\value{
Dataframe with division expenditures overall
}
\description{
Prepares dataframe on division expenditures overall for visualization in table
}
