% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{div_perstype}
\alias{div_perstype}
\title{Generate Division Expenditures by Personnel/Operating}
\usage{
div_perstype(divcode, exp_df, fy = 2023)
}
\arguments{
\item{divcode}{Division Code}

\item{exp_df}{Base process Tyler expenditures dataframe}

\item{fy}{Fiscal year preparing budget book for; default 2023}
}
\value{
Dataframe of division expenses by personnel/operating
}
\description{
Prepares dataframe of division expenditures by personnel/operating type for visualization in tables
}
