% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_preparation.R
\name{div_pull}
\alias{div_pull}
\title{Pull Division name from Division code}
\usage{
div_pull(dc)
}
\arguments{
\item{dc}{Division code}
}
\value{
Division name
}
\description{
Pulls Division name from supplied Division code
}
