% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_preparation.R
\name{find_yrstooearly}
\alias{find_yrstooearly}
\title{Identify years too early}
\usage{
find_yrstooearly(fy, bud_extract_df, tbltype = "recent")
}
\arguments{
\item{fy}{Fiscal year of this budget book}

\item{bud_extract_df}{A budget extract dataframe to pull year columns from}

\item{tbltype}{Options include "recent" (to include only preceding 3 fiscal years) and "operating" (to include preceding 4 fiscal years)}
}
\value{
String vector of year columns to exclude from dataframe
}
\description{
Pulls year columns that should be omitted from budget tables because they are before the range of years
}
