% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_tables.R
\name{flex_performancedata}
\alias{flex_performancedata}
\title{Performance Data Flextable}
\usage{
flex_performancedata(df)
}
\arguments{
\item{df}{Performance measure dataframe}
}
\value{
Flextable of performance measure dataframe
}
\description{
Create flextable out of performance measure dataframe. Ideally, should format performance measure dataframe with perf_read() beforehand.
}
