% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{gen_dept_exp_ov}
\alias{gen_dept_exp_ov}
\title{Generate Departmental Expenditures Overall}
\usage{
gen_dept_exp_ov(dept_code, exp_df, fy = 2023)
}
\arguments{
\item{dept_code}{Department code of Department to prepare dataframe for}

\item{exp_df}{Base process Tyler expenditures dataframe}

\item{fy}{Fiscal year preparing budget book for}
}
\value{

}
\description{
Generates departmental expenditures overall dataframe for visualization in tables
}
