% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{gen_dept_summary_div}
\alias{gen_dept_summary_div}
\title{Generate departmental expenses by division dataframe}
\usage{
gen_dept_summary_div(dept_code, exp_df, fy = 2023)
}
\arguments{
\item{dept_code}{Department code to generate table for}

\item{exp_df}{Base process Tyler expenditures dataframe}

\item{fy}{Fiscal year preparing budget book for}
}
\value{
Budget expenditures dataframe with factorized description field, and without too-early year columns
}
\description{
Generates departmental expenses by division dataframe for visualization in tables, grouping base processed-expenditures dataframe by department and division
}
