% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{gen_dept_summary_perstype}
\alias{gen_dept_summary_perstype}
\title{Generate Department Expenditures by Personnel or Operating Dataframe}
\usage{
gen_dept_summary_perstype(dept_code, exp_df, fy)
}
\arguments{
\item{dept_code}{Department code}

\item{exp_df}{Base process Tyler expenditures dataframe}

\item{fy}{Fiscal year preparing budget book for; default 2023}
}
\value{
Dataframe of departmental expenditures by personnel/operating expenditures for visualization in tables
}
\description{
Prepares dataframe of Departmental expenditures by operating/personnel expenditures for visualization in table
}
