% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{gen_dept_summary_rev}
\alias{gen_dept_summary_rev}
\title{Generate departmental revenue table}
\usage{
gen_dept_summary_rev(dept_code, bud_extract_rev, fy = 2023)
}
\arguments{
\item{dept_code}{Department code}

\item{bud_extract_rev}{Base budget revenue dataset}

\item{fy}{Fiscal year; default 2023}
}
\value{
Dataframe of departmental revenues for a given department
}
\description{
Generates departmental revenue table for given department
}
