% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{oneone_check}
\alias{oneone_check}
\title{One to one merge check}
\usage{
oneone_check(df1, df2, brows = F)
}
\arguments{
\item{df1}{The merged dataframe}

\item{df2}{The dataframe that was left-joined to}

\item{brows}{Whether to enter browser for debugging if there's an error; default F}
}
\value{
Nothing; stops if number of rows increased in joined dataframe and enters browser
}
\description{
Checks if number of rows increased in the joined dataframe compared to one the dataframe that was left-joined to after what was supposed to be a one-to-one merge
}
