% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_tables.R
\name{perf_read}
\alias{perf_read}
\title{Read Performance Data}
\usage{
perf_read(
  dept,
  sheet,
  .numdec = 2,
  .percsub = NULL,
  dirpath = "./data/performance/"
)
}
\arguments{
\item{dept}{Department number of Department reading performance data on.}

\item{sheet}{Name of sheet containing performance data reading in.}

\item{.numdec}{Number of decimal places; default 2.}

\item{.percsub}{Default NULL. Perf_read looks for the word "percent" in the measurement name to determine whether to round the row to 2 decimal points and multiply it by 100. In some cases "percent" does not appear in the measurement name for percentage-based metrics. In these cases, .percsub accepts a logical vector of T/F corresponding to the number of rows in the performance table, with true indicating it is a performance measure and F indicating it is not.}

\item{dirpath}{Directory performance file is stored in.}
}
\value{
Performance table formatted for visualization.
}
\description{
Reads and formats standard Departmental performance tables. Assumes performance tables saved in Workbook organized by Department titled "perf_departmentnumber.xlsx", and that each sheet of the table corresponds to one division.
}
