% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_tables.R
\name{perf_rename}
\alias{perf_rename}
\title{Rename Performance Table Columns}
\usage{
perf_rename(perfdata)
}
\arguments{
\item{perfdata}{Performance table dataframe}
}
\value{
Performance dataframe with renamed columns
}
\description{
Renames columns in standard department performance table Excel file
}
