% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_bud_pies}
\alias{plot_bud_pies}
\title{Budget Book Pie Charts}
\usage{
plot_bud_pies(df_fund, namecol, yrcol)
}
\arguments{
\item{df_fund}{Dataframe; e.g., bud_list$fund. Changing the}

\item{namecol}{Name of column to supply to label pie chart values. E.g., "accountnum" if pie chart is showing expenditures for all accounts. Input data-masked column name.}

\item{yrcol}{Year-column to visualize data for in pie chart as data-masked column.}
}
\value{

}
\description{
Create pie charts for budget book. Change the input dataframe (ideally, from the bud_list options) to change what's visualized.
}
