% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_preparation.R
\name{print_dept_exp_overall}
\alias{print_dept_exp_overall}
\title{Print Overall Department Expenditures}
\usage{
print_dept_exp_overall(dept_code, .inctable = T, .tabnum = "XX")
}
\arguments{
\item{dept_code}{The Tyler code of the department}

\item{.inctable}{Whether to include table number as prefix; true or false}

\item{.tabnum}{Number of table in document; E.g., table 12}
}
\value{
String of Overall Departmental Expenditures table name
}
\description{
Prints a standard table name for Departmental  expenditures overall tables
}
