% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_preparation.R
\name{print_div_perstype}
\alias{print_div_perstype}
\title{Print Division Expenditures by Type}
\usage{
print_div_perstype(div_code, .inctable = T, .tabnum = "XX")
}
\arguments{
\item{div_code}{The Tyler code of the division}

\item{.inctable}{Whether to include table number as prefix; true or false}

\item{.tabnum}{Number of table in document; E.g., table 12}
}
\value{
String of Overall Division Expenditures table name
}
\description{
Prints a standard name for division tables on expenditures by type
}
