% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_preparation.R
\name{print_performance}
\alias{print_performance}
\title{Print Performance Table Name}
\usage{
print_performance(deptdivname, .inctable = T, .tabnum = "XX")
}
\arguments{
\item{deptdivname}{Name of division/department as string}

\item{.inctable}{Whether to include table number as prefix; true or false}

\item{.tabnum}{Number of table in document; E.g., table 12}
}
\value{

}
\description{
Prints standard name for performance/workload measure table
}
\examples{
{
print_performance("Legal Services")
print_performance("General Government")

}
}
