% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_tables.R
\name{process_bt}
\alias{process_bt}
\title{Process budget basic table}
\usage{
process_bt(bt, colvec)
}
\arguments{
\item{bt}{Budget basic table}

\item{colvec}{Column-groups of budget basic table}
}
\value{
Basic table with values reformatted consistent with budget book formatting
}
\description{
Prepares budget pivot table converted to basic table for visual representation in flextable; intermediate function used within pt_word_table
}
