% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_tables.R
\name{pt_calc_tot}
\alias{pt_calc_tot}
\title{Funcction to calculate totals for each year}
\usage{
pt_calc_tot(pt, cols)
}
\arguments{
\item{pt}{pivot table created by pivottabler::PivotTable$new()}

\item{cols}{String vector of year columns to calculate totals for}
}
\value{
Pivot table with year columns all added in
}
\description{
Funcction to calculate totals for each year
}
