% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_tables.R
\name{pt_chng_calc}
\alias{pt_chng_calc}
\title{Add change or difference columns to pivot table}
\usage{
pt_chng_calc(pt, colname, proj, current, adpt = NULL)
}
\arguments{
\item{pt}{Pivot table to add change or difference columns to, created by pivottabler::PivotTable$new()}

\item{colname}{Name of change or difference column}

\item{proj}{Name of projected year column}

\item{current}{Name of current year column}

\item{adpt}{Name of adapted budget column; default NULL. Used to provide additional information to Departments for Departmental pivot tables}
}
\value{
Pivot table with change or difference column addded to it
}
\description{
Adds change or difference in year columns to pivot table, summing total difference/change columns and recalculating/rounding percentage columns
}
