% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_tables.R
\name{pt_createstandard}
\alias{pt_createstandard}
\title{Standard Pivot Table}
\usage{
pt_createstandard(
  df,
  rowvec = NULL,
  colvec = NULL,
  calcrow = F,
  calccoltop = F,
  deptadjust = F
)
}
\arguments{
\item{df}{Dataframe visualized in pivot table}

\item{rowvec}{Rows to visualize in pivot table; default null, so no groups appear in rows added}

\item{colvec}{Columns to visualize in pivot table; default null, so no groups appear in columns}

\item{calcrow}{whether to put year columns in rows of pivot table; default F, so years appear at top of pivot table}

\item{calccoltop}{whether to put calculation groups in columns; default F, so don't appear in columns}

\item{deptadjust}{Whether to generate pivot table with differences for adapted budget, for sharing with departments. Default false}
}
\value{
Pivot table of budget data
}
\description{
Create standard pivot table for visualizing budget data
}
