% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_tables.R
\name{pt_grps}
\alias{pt_grps}
\title{Add row or column groups}
\usage{
pt_grps(pt, cols, colgrp = T)
}
\arguments{
\item{pt}{Pivot table created by pivottabler::PivotTable$new()}

\item{cols}{Columns to add as group to row or column in pivot table; string vector}

\item{colgrp}{Whether to add columns as groups in columns of pivot table, or as rows of pivot table; default T, for columns}
}
\value{
Pivot table with columns added in as rows or columns
}
\description{
Function to add row or column groups to budget pivot table
}
