% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_tables.R
\name{pt_outputtype}
\alias{pt_outputtype}
\title{Render budget table depending on doc type}
\usage{
pt_outputtype(
  df,
  rowvec = NULL,
  colvec = NULL,
  calcrow = F,
  .ptsavename = NULL,
  .caption = NULL,
  deptadjust = F
)
}
\arguments{
\item{df}{Budget dataframe to represent in budget table}

\item{rowvec}{String column-names of the dataframe to add as row-groups to the pivot table; default NULL, for none}

\item{colvec}{String columnn-names of the dataframe to add as column-groups to the pivot table; default NULL, for none}

\item{calcrow}{Whether to display year columns of budget dataframe as rows of the pivot table; default F, so shown as columns}

\item{.ptsavename}{Filename to save pivot table as in Excel document; default NULL, so not saved}

\item{.caption}{Caption for pivot table; default NULL, so no caption}

\item{deptadjust}{Whether including adotped budget information and comparisons for sharing of table with departments; default F, so adopted budget information excluded}
}
\value{
pivottabler pivot-table of budget data or flextable pivot-table of budget data
}
\description{
Renders either a pivottabler pivot-table or flextable pivot-table based on type of document; the former in Latex documents, the latter otherwise
}
