% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_tables.R
\name{pt_render_dept_divcosts}
\alias{pt_render_dept_divcosts}
\title{Pivot Table of Departmental Expenditures by Division}
\usage{
pt_render_dept_divcosts(dept_code, bud_dept_list, fy = 2023, deptadjust = F)
}
\arguments{
\item{dept_code}{Code of department creating table for}

\item{bud_dept_list}{List of grouped budget expenditures}

\item{fy}{Fiscal year for this budget book}

\item{deptadjust}{Whether including adotped budget information and comparisons for sharing of table with departments; default F, so adopted budget information excluded}
}
\value{
Pivot table of departmental expenditures by divisions
}
\description{
Create and render pivot table of departmental expenditures by division
}
