% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_tables.R
\name{pt_render_dept_overall}
\alias{pt_render_dept_overall}
\title{Create and Render Department Revenue Pivot}
\usage{
pt_render_dept_overall(dept_code, bud_extract_rev, deptadjust, fy = 2023)
}
\arguments{
\item{dept_code}{Code of department to show revenue for}

\item{bud_extract_rev}{Budget extract revenue file}

\item{deptadjust}{Whether including adotped budget information and comparisons for sharing of table with departments; default F, so adopted budget information excluded}

\item{fy}{Fiscal year for this budget}
}
\value{
Rendered pivot table of departmental revenues by source
}
\description{
Creates/renders departmental revenue for given department in pivot table
}
