% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_tables.R
\name{pt_render_dept_perstype}
\alias{pt_render_dept_perstype}
\title{Pivot table of Departmental Expenditures by Personnel/Operating}
\usage{
pt_render_dept_perstype(deptcode, bud_dept_list, fy = 2023, deptadjust = F)
}
\arguments{
\item{deptcode}{Department code}

\item{bud_dept_list}{List of grouped budget expenditures}

\item{fy}{Fiscal year for this budget book}

\item{deptadjust}{Whether including adotped budget information and comparisons for sharing of table with departments; default F, so adopted budget information excluded}
}
\value{
Pivot table of departmental expenditures by personnel/operating type
}
\description{
Generate pivot table of departmental expenditures by personnel/operating type
}
