% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_tables.R
\name{pt_rendercreate}
\alias{pt_rendercreate}
\title{Create and Render Budget Pivot}
\usage{
pt_rendercreate(
  df,
  rowvec = NULL,
  colvec = NULL,
  calcrow = F,
  .ptsavename = NULL,
  deptadjust = F
)
}
\arguments{
\item{df}{Dataframe to represent in budget pivot table}

\item{rowvec}{Columns of dataframe to add as row-groups to budget pivot table; default NULL, for none}

\item{colvec}{Columns of dataframe to add as column-groups to budget pivot table; default NULL, for none}

\item{calcrow}{Whether to display year columns as rows of budget pivot table; default F}

\item{.ptsavename}{Filename of outputted Excel document that contains pivot table; default NULL, in which case pivot table not saved as Excel doc}

\item{deptadjust}{Whether you're generating pivot tables to share with departments that include comparisons with the adopted budget book; default F, so no adjustment made}
}
\value{
Rendered pivot table of budget data
}
\description{
Creates and renders budget pivot table in one step
}
