% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_tables.R
\name{pt_save}
\alias{pt_save}
\title{Save budget pivot table}
\usage{
pt_save(pt, filename, deptadjust = F)
}
\arguments{
\item{pt}{Pivot table to save}

\item{filename}{Filename of Excel document, don't include .xlsx}

\item{deptadjust}{Whetherto save pivot tables to folder for sharing adopted budget information with departments; default false}
}
\value{
Outputted Excel document with pivot table
}
\description{
Saves budget pivot table in Excel workbook
}
