% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_tables.R
\name{pt_style}
\alias{pt_style}
\title{Style budget pivot table}
\usage{
pt_style(pt, df, rowvec = NULL, colvec = NULL)
}
\arguments{
\item{pt}{Budget pivot table to style}

\item{df}{Budget dataframe pivot table representing}

\item{rowvec}{Row-groups of budget pivot table, represented as string vector}

\item{colvec}{Column-groups of budget pivot table, represented as string vector}
}
\value{
Pivot table with standard styling for budget book
}
\description{
Applies prescribed style to budget pivot table
}
