% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_tables.R
\name{pt_word_table}
\alias{pt_word_table}
\title{Create Standard Flextable in Word}
\usage{
pt_word_table(
  df,
  rowvec = NULL,
  colvec = NULL,
  calcrow = F,
  .ptsavename = NULL,
  .caption = NULL,
  deptadjust = F
)
}
\arguments{
\item{df}{Budget dataframe to represent in flextable}

\item{rowvec}{Columns of budget dataframe to add as groups in flextable; default NULL, for none}

\item{colvec}{Columns of budget dataframe to add as columns in flextable; default NULL, for none}

\item{calcrow}{Whether to represent year columns in rows of flextable; default F, for not}

\item{.ptsavename}{Name of file to save pivot table as Excel document; default NULL, so not saved}

\item{.caption}{Caption of flextable; default NULL, so no caption added}

\item{deptadjust}{Whether including adotped budget information and comparisons for sharing of table with departments; default F, so adopted budget information excluded}
}
\value{
Flextable of budget data styled in standard format
}
\description{
Creates standard budget flextable for output to Word documents
}
