% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_tables.R
\name{pt_year_calc_loop}
\alias{pt_year_calc_loop}
\title{Add all change or difference columns to pivot table}
\usage{
pt_year_calc_loop(pt, cols, proj, current, adpt = NULL)
}
\arguments{
\item{pt}{pivot table created by pivottabler::PivotTable$new()}

\item{cols}{String vector of change or difference columns in budget dataframe being visualized in pivot table}

\item{proj}{Name of projected year column}

\item{current}{Name of current FY column}

\item{adpt}{Name of adapted budget column; default NULL. Used to provide additional information to Departments for Departmental pivot tables. Default NULL, so excluded}
}
\value{
Budget pivot table with change and difference columns added in
}
\description{
Function to add all change or difference columns in budget dataframe to pivot table created by pivotTabler::PivotTable$new()
}
