% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_prep.R
\name{quick_datatbl}
\alias{quick_datatbl}
\title{Quick Excel Datatable}
\usage{
quick_datatbl(df, fileend, directory = "./data/output/")
}
\arguments{
\item{df}{Dataframe to save}

\item{fileend}{String name of Excel file (omit .xlsx)}

\item{directory}{Directory to save file in; default ./data/output}
}
\value{
Nothing, but saves dataframe as Excel workbook
}
\description{
Saves a dataframe as a quick, formatted Excel datatable in one step
}
