% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{recode_personnel_other}
\alias{recode_personnel_other}
\title{Recode expenditures as personnel, operating, or non-departmental
Code expenditures as perosonnel, operating, or non-departmental types}
\usage{
recode_personnel_other(
  df,
  category_num_col = category_num,
  division_col = division,
  div_desc = division_desc,
  accountnum_desc_col = accountnum_desc
)
}
\arguments{
\item{df}{Ungrouped Tyler-extract report, ideally processed by base_report_extract}

\item{category_num_col}{Data-masked column identifying the category number; default category_num}

\item{division_col}{Data-masked column identifying the full division; defaults division}

\item{div_desc}{Data-masked column identifying description of division; defaults division_desc}

\item{accountnum_desc_col}{Data-masked column identifying the account number; defaults accountnum_desc}
}
\value{
Dataframe with perstype column of factor type categorizing expenditures as "Personnel" or "Other Operating Expenses"
}
\description{
Recode expenditures as personnel, operating, or non-departmental
Code expenditures as perosonnel, operating, or non-departmental types
}
