% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{rename_yeartoyear}
\alias{rename_yeartoyear}
\title{Rename year-change columns
Rename columns representing changes from one year to another year, and difference columns from adopted to projected to names for representation in tables and visualizations}
\usage{
rename_yeartoyear(vec)
}
\arguments{
\item{vec}{Character vector of change and difference column names}
}
\value{
Vector of renamed columns
}
\description{
Rename year-change columns
Rename columns representing changes from one year to another year, and difference columns from adopted to projected to names for representation in tables and visualizations
}
