% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{revenue_factorize}
\alias{revenue_factorize}
\title{Factorize Revenue/Expenditure Column
Turns revenue/expenditure column of Tyler exract report dataframe into factor in order of revenue, then expense}
\usage{
revenue_factorize(df, colname = account.type)
}
\arguments{
\item{df}{Tyler Extract report dataframe with a column categorizing accounts as "Revenue" or "Expense"}

\item{colname}{Data-masked name of column categorizing accounts as "Revenue" or "Expense"}
}
\value{

}
\description{
Factorize Revenue/Expenditure Column
Turns revenue/expenditure column of Tyler exract report dataframe into factor in order of revenue, then expense
}
\examples{
{
df <- data.frame(account.type = c("Revenue", "Expense", "Revenue"),
account_desc = c("Taxes", "Maintenance", "Parking fees"))

revenue_factorize(df)

df <- data.frame(revcol = c("Revenue", "Expense", "Revenue"),
account_desc = c("Taxes", "Maintenance", "Parking fees"))

revenue_factorize(df, revcol)

}
}
