% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{sep_budget}
\alias{sep_budget}
\title{Separate budget columns
Creates a number column and description column from standard Tyler Extract report column that's a combination of a number and description. E.g., Departments are contained in the "department" column and represented like 1000 Department Name; this creates a department_num column of 1000, and a 'department_desc" column of "Department Name."}
\usage{
sep_budget(df, colname)
}
\arguments{
\item{df}{Minimally-processed Tyler Extract report dataframe}

\item{colname}{String name of column with a numeric part and string part separated by a space}
}
\value{
Dataframe with original column, and column for number-part of original column and string part
}
\description{
Separate budget columns
Creates a number column and description column from standard Tyler Extract report column that's a combination of a number and description. E.g., Departments are contained in the "department" column and represented like 1000 Department Name; this creates a department_num column of 1000, and a 'department_desc" column of "Department Name."
}
\examples{
{
c <- data.frame(sepcolumn = c("52 Cats", "35 Dogs", "42 Fish"))

sep_budget(c, "sepcolumn")
}

}
