% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_functs.R
\name{standard_knitr}
\alias{standard_knitr}
\title{Standard knit settings
Sets default knitting options for RMarkdown budget documents}
\usage{
standard_knitr()
}
\value{
Default chunk options for RMarkdown
}
\description{
Standard knit settings
Sets default knitting options for RMarkdown budget documents
}
