% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_accounts_complete}
\alias{table_accounts_complete}
\title{Complete Table Accounts}
\format{
A data frame with 331 rows and 6 variables:
\describe{
\item{account.name }{Account name in Tyler}
\item{accountnum }{Combined account number and account description; distinct from account.name in some instances}
\item{accountnum_num }{An account number associated with a budget category}
\item{budget_cat_overall}{A category used in the City's budget tables}
\item{budget_cat_overall_desc}{Description of budget category used in City tables}
\item{budget_cat_overall_num }{Number of budget category used in City tables}
}
}
\source{
{City staff}
}
\usage{
table_accounts_complete
}
\description{
Dataframe of all budget accounts and their associated categories.
}
\keyword{datasets}
