% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_tables.R
\name{transform_num}
\alias{transform_num}
\title{Transform Numeric Performance Measures}
\usage{
transform_num(df, colname, .numdec = 2, .percsub = NULL)
}
\arguments{
\item{df}{Performance measure dataframe.}

\item{colname}{Name of performance measure column to format as number; string.}

\item{.numdec}{Number of decimal places; default 2.}

\item{.percsub}{Default NULL. Perf_read looks for the word "percent" in the measurement name to determine whether to round the row to 2 decimal points and multiply it by 100. In some cases "percent" does not appear in the measurement name for percentage-based metrics. In these cases, .percsub accepts a logical vector of T/F corresponding to the number of rows in the performance table, with true indicating it is a performance measure and F indicating it is not.}
}
\value{
Formatted dataframe with formatted performance measure column
}
\description{
Formats numeric performance measure columns for representation in tables.
}
