% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_prep.R
\name{write_output}
\alias{write_output}
\title{Write Dataframe to Excel}
\usage{
write_output(df, filename)
}
\arguments{
\item{df}{Dataframe to write to an Excel table}

\item{filename}{Name of output file, including directory}
}
\value{
Outputs Excel document to filename
}
\description{
Write dataframe to Excel formatted as a table
}
\examples{
{
write_output(mtcars, "./mtcars.xlsx")
}
}
