% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{year_transform}
\alias{year_transform}
\title{Rename Year Columns
Renames year columns to match budget book descriptions, for representation in visualizations and tables}
\usage{
year_transform(yrcol)
}
\arguments{
\item{yrcol}{String or string-vector associated with year column names from tpbudget::base_process_extract()}
}
\value{
String representation of year column
}
\description{
Rename Year Columns
Renames year columns to match budget book descriptions, for representation in visualizations and tables
}
\examples{
{

year_transform("yr2021_2022.total.budget")

year_transform(c("yr2021_2022.ytd.activity.through.april", "yr2021_2022.total.budget"))

}
}
